requirements.txt
# Core requirements
numpy>=1.22.0
pandas>=1.4.0
matplotlib>=3.5.0
streamlit>=1.10.0
plotly>=5.8.0
pyyaml>=6.0
requests>=2.27.0
python-dotenv>=0.19.0

# Blockchain-specific libraries
web3>=5.29.0
solana>=0.25.0
cosmos-sdk>=0.1.0
pycoingecko>=2.2.0

# Data analysis
scikit-learn>=1.0.0
statsmodels>=0.13.0
scipy>=1.8.0

# Storage
sqlalchemy>=1.4.0
pymongo>=4.1.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.1.0
flake8>=4.0.0
isort>=5.10.0